Conflicts:
	Sources/Athena/Athena_Environment.h
	Sources/Athena/Athena_FileIO.cpp
	Sources/Athena/Athena_Global.h
	Sources/Athena/Athena_Instance.h
	Sources/Athena/Athena_Resource.h
	Sources/Include/Athena_Types.h

+++ Sources/Include/Athena_Types.h
@@ -41,9 +41,9 @@ namespace ATHENA
 	typedef signed char aalSByte;
 	typedef unsigned short aalUWord;
 	typedef signed short aalSWord;
-	typedef unsigned long aalULong;
-	typedef signed long aalSLong;
-
+	typedef unsigned int aalULong;
+	typedef signed int aalSLong;
+ 
 	typedef signed aalSInt;
 	typedef float aalFloat;
